2. basic_search_agent/pyproject.toml
[project]
name = "2-basic-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "isort>=7.0.0",
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.19",
]
